Move common database functions to other file
